added new features such as : index.html or regiester.html
