+++ flask_api.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+app = Flask(__name__)
+app.secret_key = "secret_key"
