+++ requirements.txt
@@ -1,3 +1,4 @@
 flask==3.1.0
 psycopg2==2.9.10
 Flask-RESTful==0.3.10
+flask-smorest
