+++ templates/login_new.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="ro">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="{{ url_for('static', filename='style/login_new.css') }}">
+    <title>Autentificare - BiblioTech</title>
+
+    <!-- Font Awesome pentru iconițe -->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
+
+</head>
+<body>
+    <div class="login-container">
+        <div class="logo-container">
+            <a href="index.html" class="logo">
+                <i class="fas fa-book-open"></i>
+                BiblioTech
+            </a>
+        </div>
+        <h1>Autentificare</h1>
+        <p class="subtitle">Introduceți datele pentru a vă conecta la contul dvs.</p>
+
+        <form action="#" method="post">
+            <div class="form-group">
+                <label for="username">Username</label>
+                <input type="text" id="username" name="username" placeholder="nusername_ak27" required>
+            </div>
+            <div class="form-group">
+                <label for="password">Parolă</label>
+                <input type="password" id="password" name="password" required>
+                <a href="#" class="forgot-password">Ai uitat parola?</a>
+            </div>
+            <button type="submit" class="btn">Autentificare</button>
+        </form>
+
+        <div class="register-link">
+            Nu ai un cont? <a href="register.html">Înregistrează-te</a>
+        </div>
+    </div>
+
+    <a href="{{ url_for('welcomes') }}" class="back-link">
+        <i class="fas fa-arrow-left"></i>
+        Înapoi la pagina principală
+    </a>
+</body>
+</html>
