+++ templates/user.html
@@ -17,7 +17,7 @@
     </div>
 
     <nav id="toolbar" class="toolbar">
-    <button id="all-books-btn" class="btn" onclick="window.location.href='/user_view_books'">All Books</button>
+    <button id="all-books-btn" class="btn" onclick="window.location.href='/view_books'">All Books</button>
     <button id="pending-books-btn" class="btn" onclick="getPendingBooks()">Pending Books</button>
     <button id="borrow-book-btn" class="btn" onclick="borrowBook()">Borrow a Book</button>
     <button id="return-book-btn" class="btn" onclick="returnBook()">Return a Book</button>
