+++ templates/view_books.html
@@ -7,10 +7,17 @@
 </head>
 <body>
     <header>
-        <h1>Library Books</h1>
-        <a href="{{ url_for('web_home') }}">
-            <button class="admin-btn">Back to Admin Page</button>
-        </a>
+        {% if session['is_admin']== 'Da' %}
+            <h1>Library Books</h1>
+            <a href="{{ url_for('web_home') }}">
+                <button class="admin-btn">Back to Admin Page</button>
+            </a>
+        {% else %}
+            <h1>Library Books</h1>
+            <a href="{{ url_for('welcomes') }}">
+                <button class="admin-btn">Back to User Page</button>
+            </a>
+        {% endif %}
     </header>
 
     <main>
