+++ templates/view_books_updated.html
@@ -1,0 +1,85 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Admin - View Books</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='style/admin.css') }}">
+</head>
+<body>
+
+<h1>All Books</h1>
+
+<table border="1">
+    <thead>
+        <tr>
+            <th>Nr.</th>
+            <th>Title</th>
+            <th>Author</th>
+            <th>Description</th>
+            <th>Page Count</th>
+            <th>Actions</th>
+        </tr>
+    </thead>
+    <tbody>
+        {% for book in books %}
+        <tr id="row-{{ book.book_id }}">
+            <td>{{ loop.index }}</td>
+            <td id="title-{{ book.book_id }}">{{ book.title }}</td>
+            <td id="author-{{ book.book_id }}">{{ book.author }}</td>
+            <td id="description-{{ book.book_id }}">{{ book.description }}</td>
+            <td id="pagecount-{{ book.book_id }}">{{ book.page_count }}</td>
+            <td>
+                <form action="{{ url_for('delete_book', book_id=book.book_id) }}" method="post" style="display:inline;">
+                    <button type="submit">Delete</button>
+                </form>
+<!--                <button onclick="deleteBook({{ book.book_id }})">Delete</button>-->
+
+                <button onclick="showEditForm({{ book.book_id }})">Edit</button>
+            </td>
+        </tr>
+
+        <!-- Formularul de editare ascuns -->
+        <tr id="edit-row-{{ book.book_id }}" style="display: none;">
+            <td colspan="6">
+                <form method="post" action="{{ url_for('edit_book', book_id=book.book_id) }}" onsubmit="return confirm('Save changes?');">
+                    Title: <input type="text" name="title" value="{{ book.title }}">
+                    Author: <input type="text" name="author" value="{{ book.author }}">
+                    Description: <input type="text" name="description" value="{{ book.description }}">
+                    Pages: <input type="number" name="page_count" value="{{ book.page_count }}">
+                    <button type="submit">Save</button>
+                    <button type="button" onclick="cancelEdit({{ book.book_id }})">Cancel</button>
+                </form>
+            </td>
+        </tr>
+        {% endfor %}
+    </tbody>
+</table>
+
+<script>
+function showEditForm(bookId) {
+    document.getElementById(`edit-row-${bookId}`).style.display = '';
+}
+
+function deleteBook(bookId) {
+    if (!confirm("Are you sure you want to delete this book?")) return;
+
+    fetch(`/delete_book/${bookId}`, {
+        method: 'DELETE'
+    })
+    .then(response => {
+        if (response.ok) {
+            alert("Book deleted!");
+            location.reload();
+        } else {
+            alert("Error deleting the book.");
+        }
+    });
+}
+
+function cancelEdit(bookId) {
+    document.getElementById(`edit-row-${bookId}`).style.display = 'none';
+}
+</script>
+
+</body>
+</html>
